refactor: 예약 번호의 타입 수정 및 번호 자동 증가 책임 repository로 이동
