+++ docs/README.md
@@ -6,4 +6,3 @@
 - [x] 공통되는 url requestMapping으로 빼기 (Q. 이렇게 쓰는 게 맞나?)
 - [x] repository의 db 경우 고려 interface 분리 
 - [x] 시간 출력 방식 요구사항에 맞게 수정
-- [ ] TODO 제거
