+++ src/main/java/roomescape/db/domain/Reservation.java
@@ -1,23 +1,23 @@
-package roomescape.domain;
+package roomescape.db.domain;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
 
 public class Reservation {
 
-    private final Long id;
+    private final long id;
     private final String name;
     private final LocalDate date;
     private final LocalTime time;
 
-    public Reservation(final Long id, final String name, final LocalDate date, final LocalTime time) {
+    public Reservation(final long id, final String name, final LocalDate date, final LocalTime time) {
         this.id = id;
         this.name = name;
         this.date = date;
         this.time = time;
     }
 
-    public Long getId() {
+    public long getId() {
         return id;
     }
