+++ src/main/java/roomescape/service/ReservationService.java
@@ -1,21 +1,18 @@
 package roomescape.service;
 
 import java.util.List;
-import java.util.concurrent.atomic.AtomicLong;
-import roomescape.domain.Reservation;
-import roomescape.dto.ReservationRequest;
-import roomescape.dto.ReservationResponse;
-import roomescape.repository.InMemoryReservationRepository;
-import roomescape.repository.ReservationRepository;
+import roomescape.db.domain.Reservation;
+import roomescape.web.dto.ReservationRequest;
+import roomescape.web.dto.ReservationResponse;
+import roomescape.db.repository.InMemoryReservationRepository;
+import roomescape.db.repository.ReservationRepository;
 
 public class ReservationService {
 
     private final ReservationRepository reservationRepository;
-    private final AtomicLong index;
 
     public ReservationService() {
         this.reservationRepository = new InMemoryReservationRepository();
-        this.index = new AtomicLong(1);
     }
 
     public List<ReservationResponse> read() {
@@ -24,13 +21,14 @@ public class ReservationService {
     }
 
     public ReservationResponse add(final ReservationRequest reservationRequest) {
-        Long id = index.getAndIncrement(); // TODO: 여기 있어도 되는 로직일까?
+        long id = reservationRepository.getCurrentId();
         Reservation reservation = reservationRequest.toReservation(id);
+
         reservationRepository.save(reservation);
         return ReservationResponse.from(reservation);
     }
 
-    public void remove(final Long id) {
+    public void remove(final long id) {
         reservationRepository.delete(id);
     }
 }
