+++ src/main/java/roomescape/web/dto/ReservationRequest.java
@@ -1,12 +1,12 @@
-package roomescape.dto;
+package roomescape.web.dto;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
-import roomescape.domain.Reservation;
+import roomescape.db.domain.Reservation;
 
 public record ReservationRequest(String name, LocalDate date, LocalTime time) {
 
-    public Reservation toReservation(final Long id) {
+    public Reservation toReservation(final long id) {
        return new Reservation(id, name, date, time);
     }
 }
