+++ src/main/java/roomescape/web/dto/ReservationResponse.java
@@ -1,12 +1,12 @@
-package roomescape.dto;
+package roomescape.web.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.List;
-import roomescape.domain.Reservation;
+import roomescape.db.domain.Reservation;
 
-public record ReservationResponse(Long id, String name, LocalDate date, @JsonFormat(pattern = "HH:mm") LocalTime time) {
+public record ReservationResponse(long id, String name, LocalDate date, @JsonFormat(pattern = "HH:mm") LocalTime time) {
 
     public static ReservationResponse from(final Reservation reservation) {
         return new ReservationResponse(reservation.getId(), reservation.getName(), reservation.getDate(),
